mgsim: fixed place inheritance bug and group create bug
The first bug was that default creates did not work as expected. A default create should copy the place from the parent, even though that place may 
have been optimized to something else (i.e., copy the logical place). For instance, a group create may be forced to local if there's not enough 
threads. However, default creates made by any of those threads must still be group creates, not local creates. This was not the case and has been 
fixed.

The second bug was for group creates. When such a create is sent along the ring, the parent processor receives the creates and updates the previous 
processor's family immediately, as opposed to next cycle (on non-parent processors). Because of the split-phase sub-cycle implementation, the actual 
family on this previous processor was not fully instantiated yet, so an assertion was triggered. This was fixed by recognizing that the create has 
gone all the way around on the previous processor, and settings the family entry field as part of the regular allocation of the entry.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@1933 e97e5017-a994-416e-809d-76780e9f78db