mgsim: Fixed the infinite waiting pipeline bug
The bug was caused by having threads use a global that was written after the family was created. The threads would start suspending on the empty global, and write back a Waiting register. When the global was finally read from memory, threads writing back Waiting would be woken up as usual. However, because the threads followed each other closely in the pipeline, each would use the pipeline bypasses to pick up the Waiting register from the previous thread, further in the pipeline. Thus, none of the threads would pick up the Full register from the register file, and they would all keep getting rescheduled when they write back their register.

The fix was to have the Read Stage of the pipeline check all inputs and determine the correct final value from those, and not just grab the first one that matched. A slight rewrite of the Read Stage was required for this.

A regression test has also been added for this, and the 'make check' scripts adjusted to be able to timeout on tests.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@1998 e97e5017-a994-416e-809d-76780e9f78db