mgsim: Implemented exclusive creates.
After restructuring of the create logic, a second create queue has been added. On any core, if a create has been marked as exclusive, it will be queued on the exclusive create queue. A register will keep track of the currently active exclusive family on that core.

Note that this means that group creates must be placed on the same core in order to be exclusive, even if they run on the same ring. As such, exclusive creates can only be local or delegated for deterministic exclusiveness.

Also, 'make check' has been altered to use the debug version instead of the release version and some better error checking to mark assertions as FAILs.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2041 e97e5017-a994-416e-809d-76780e9f78db