mgsim: fixed an assertion bug with global requests.
Sometimes a remote global request would trigger an assertion error because the family was already cleaned up so the request could not be propagated. This was fixed by adding a reverse cleanup token to every family, in addition to the already existing forward synchronization token. The parent processor starts with the cleanup token. A processor forwards the token if (it has it and) all threads have localy terminated, so no more registers can be requested.

In developing this fix, another issue came to light: It is possible for a group create broadcast to be delayed after receiving the token, because the send buffer is full and stalled. During this delay, the token is sent along, violating the create token protocol. The solution to this problem was twofold: first, a "token used" flag was added. Upon requesting the token for a create this flag is cleared, and set once a create is put in the send buffer. The network logic has also been reshuffled to put create broadcasts and token transmission in the same channel. So as long as there is a create to send, the token will not be processed.

Finally, in order for more error checking, all network registers and buffers are now inherently arbitrated and all allowed sources have to be explicitly set. This was done after realising that several processes could want to send data via the same unarbitrated network buffer in the same cycle. This would've likely asserted the simulator, as one process would not be able to continue when it thought it would be.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2089 e97e5017-a994-416e-809d-76780e9f78db