mgsim: implemented new internal configuration system
The new system no longer fills predefined structures for each component for configuration data. Now it passes the configuration data store to every component and it will grab its own configuration data from it. This allows different component types without the main system requiring any knowledge of the details of the component.

Along with this, the memory type is now a run-time option (instead of compile-time), which can be overriden on the command line. The current built-in memories are Ideal, Parallel, Banked and RandomBanked.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2164 e97e5017-a994-416e-809d-76780e9f78db