mgsim: implemented FPU sharing
The configuration has a new option, NumProcessorsPerFPU, that will determine the number of CPUs connected to a single FPU. Each CPU has a single input queue and the FPU services operation from the longest queue. The internals of the FPU has also been changed. Instead of having a pipeline for all operations with the same latency, the FPU now has a non-pipelined execution unit for every FP operation. A new operation will not be started until the execution unit has been cleared by writing back the result to the appropriate register file.

Note that simulator will throw an error if the number of processors in a place does not divide by the number of processors per FPU. This has been done to match probably hardware implementations, and to remove imbalances for certain FPUs.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2174 e97e5017-a994-416e-809d-76780e9f78db