mgsim: fixed a pipeline bug and a compilation warning.
The pipeline bug occurs when a long latency operation is started on a shared that is arelady being waited on. The register is waiting and the written value is empty. When the thread then waits on the result of the operation itself, it assumes the empty value completely overwrites the waiting value; this is not the case. The read stage has been update to account for this possibility.

The warning was a simple cast problem for printf operands.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2190 e97e5017-a994-416e-809d-76780e9f78db