mgsim: fixed a read-from-shared bug.
This bug occurs when reading from a shared that's the result of a long-latency operation. The decode stage fills the remote_reg information since the 
shared would normally have to be transmitted when read. However, because the read fails, and the thread needs to suspends, the read logic sees the 
value remote_reg information and sends a remote request, which shouldn't happen for shareds.

The solution was to not fill out the remote_reg information unless the shared register is the destination operand in the instruction.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2191 e97e5017-a994-416e-809d-76780e9f78db