mgsim: implemented parallel FPU operation dispatch
Previously, only one FP operation per cycle was taken from the input queues and put on a pipeline. Since this is against the point of FPU sharing and multiple pipelines, this was altered so every input queue now tries to put their data onto a pipeline in parallel.

Arbitrators have been placed on the pipelines. The mapping between inputs and pipelines in case there are multiple pipelines that can accept an operation type is a modulo mapping.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2207 e97e5017-a994-416e-809d-76780e9f78db