mgsim: fixed Sparc multi-register parent-shared bug
When a multi-register shared was used was the result of a long-latency operation (such as 64-bit floating point operation on the Sparc), only one 
register would get forwarded to the next CPU on the delayed writeback, instead of all of them. This would only happen if the last thread in a family 
did not run on the parent core. Normally, the first register would wake up the dependent thread who would then request the other registers. With 
parent shareds, no registers are requested, so all must be explicitely sent.

The problem was that the waiting state would only be written to one register. This was changed to write all registers to Empty on the long-latency 
dispatch, and store the remote information in all of them. The memory and waiting threads, if any, would only be stored in the first register to 
avoid waking up threads multiple times.

A regression test has been added for this scenario.


Additionally, the store value register address for stores on the Sparc have been put in a separate pipeline register, because putting it in the 
destination register just confuses the decode stage which will assume the destination register is written (which, for a store, it is not).



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2475 e97e5017-a994-416e-809d-76780e9f78db