mgsim: fixed an simulator-hang bug and an assert bug
The simulator-hang bug occurred during resource deadlock on a core, with a pending group create on the other core. The resource deadlock prevented the create from completing, and thus prevent the token from being sent to the other core. However, that core's create process was in the 'waiting for token' state and kept running SUCCESS every cycle, thus tricking the kernel into thinking it was still doing something and preventing it from recognizing deadlock. The solution was to make the create process return FAILED on the two states where it was just waiting for something to complete (token and cache-line load).

The assert bug occured in the Writeback stage, where a write caused a thread wakeup. If the instruction that did the write itself switched as well, there would be two thread activations at the same time, which caused an assertion. The solution was to recognize in the Writeback stage that a write would wakeup a thread (the old register state is known, so this is trivial), and delay the reschedule of the thread by a cycle.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2477 e97e5017-a994-416e-809d-76780e9f78db