mgsim: fixed deadlock and added sanity check on instructions
A deadlock situation occured because the priorities from two exclusive components were not specified in the right order. These have been re-ordered.

Also, due to 'weird' behavior in a recent program, which was caused by a memory load to a dependent, a sanity check was addedd to the Decode stage to throw an exception when a program attempts to use a dependent register as the target of an instruction.

The maximum concurrent pushes for buffers has also been increased because such a situation surfaced.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2558 e97e5017-a994-416e-809d-76780e9f78db