mgsim: added the ability to load legacy files
As a simple addition to the simulator, it can now load legacy ELF files for 
the architecture besides the new microthreaded ELF files. Depending on the 
value of the architceture field in the ELF file, the initial family will be 
legacy or not.

However, note that plainly loading legacy ELF binaries has little immediate 
benefit, as the calling convention is not set up properly and, most likely, 
the entry point function would try to 'return' instead of terminate the thread.

However, if a small legacy wrapper can be created that, for instance, 
sets up the stack and HALTs after executing the main thread, legacy programs 
with no external dependencies (such as syscalls) can be executed.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2689 e97e5017-a994-416e-809d-76780e9f78db