mgsim: implemented hardware part of resource deadlock prevention
As outlined in note svp28, the deadlock prevention protocol requires several hardware adjustments. The following adjustments have been made in the simulator:

A global context-available and reserve-context signal have been introduced to allow group creates to be executed locally if no context can be allocated on all cores on the group. For this feature to work, the sensitivity of all storage classes was reworked. No longer is the sensitive process specified in the constructor. A sensitive process must be registered via a method on the storage class, thus allowing multiple processes to be sensitive on a single storage instance.

The Family Table, Thread Table and Register Allocator can now deal with normal, reserved and exclusive contexts. A single exclusive context is always used for exclusive creates.

To improve performance of group creates, the token protocol has also been optimized by adding a global want-token signal. If set, at least one core wants the token and the token owner will send it to the next core when it is no longer being used. No longer will a request for the token have to proceed along the ring.

Delegations will now fail if no context or the last context is acquires on the remote core. The failure sends back a message to the parent who will then re-initialize the proxy entry for a local create. A delegation flood test case was added to test this.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2773 e97e5017-a994-416e-809d-76780e9f78db