mgsim: fixed physical block size bug
During family parameter sanitization, the physical block size for independent families would be set to the originally specified block size by the program, based on the idea that the virtual block size would be an even distribution of the load among the cores, and the specified block size would provide an upper limit on threads. However, this should only be done if it constrains the virtual block size (i.e. number of threads). Since it's possible to specify a block size much larger than the number of threads in the family, it would be used to allocate that many registers, depleting resources unnecessarily. This has been fixed so the block size for independent families is only used as the physical block size if it less than calculated virtual block size.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2920 e97e5017-a994-416e-809d-76780e9f78db