mgsim: fixed some bugs in example programs
The fibonacci program uses register $2 to pass the problem size from the environment to the program. However, since r2111 register $2 is overwritten with a pointer to the static configuration area. Fibonacco now uses $10 for the problem size, like other programs.

Several programs on the Sparc version used the JMP (Jump and Link) opcode instead of BA (Branch Always) for branches, which is the wrong opcode and produces invalid code when the program base address is too high. The JMPs have been modified to BAs.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2960 e97e5017-a994-416e-809d-76780e9f78db