[mgsim] fixed a bug with the checks when writing to a register
The register file performs checks to make sure writes to registers are sane and do not mess up certain states. One of these is non-memory components overwriting registers that are currently being loaded from memory. However, when remote requests for registers arrive, this information is stored alongside the memory information. Thus, a non-memory component updates (writes to) a register being loaded from memory and would trigger an exception. Since this is actually a valid scenario, the code that checks the writes has been updated to allow these kinds of writes. It nows checks that the memory information isn't actually changed, unless the memory writes it.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3390 e97e5017-a994-416e-809d-76780e9f78db