[mgsim] Added branch optimalisation
We want to be able to get the current program counter in the Alpha architecture with a "br $pv, 0" instruction, which jumps to the next instruction, and stores that PC in the designated register.

To avoid a forced pipeline flush in this case, an exception was added in the pipeline that ignores the PC change if the target is the next instruction anyway.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3629 e97e5017-a994-416e-809d-76780e9f78db