[mgsim] Major rewrite of family creation mechanism to support continuation creates and simplify the implementation and compiler development
The old create system took the created family's globals and shareds directly from the context of the parent thread. This 'pull' model had two downsides:
* It created a dependency on the parent thread's context that was undesirable for continuation creates.
* It complicated the compiler design, as the registers that were shared with the child family would be off-limit between the create and sync.

To solve these two issues, the 'pull' model was changed to a 'push' model, where the parent thread would explicitely copy the globals and shareds over to the child family's context, thus removing any dependency on its context. Next to this, there was one dependency left on the parent context: the sync register. This register was cleared by create and written when the family has completed. This meant that the register was still off-limits between create and sync in the program. To remove this last dependency, the create instruction was split into to; 'create' and 'sync'.
The destination register of the new 'create' is cleared as usual but is written when the family's context has been created and globals and shareds can be written to it. The 'sync' instruction writes its destination register when the specified family has completed. Since both these instructions can be considered atomic by the compiler, no special register allocation logic is required between the create and sync point in a program.

The write globals and shareds to a family context, two new instructions have been added: 'putg' and 'puts'. To read the last shared back after the sync, the instruction 'gets' has been added.

To support continuation creates, a 'detach' instruction has been added that signals to the specified family that the parent thread no longer requires access to its context. This can be used after the sync or after getting the last shared in non-continuation creates or after sending the last global or shared in a continuation create.


These changes significantly affected the internal operation of the simulator. The bi-directional ring network is now uni-directional as shareds and globals are always pushed and all contexts are guaranteed to be created on all cores. Reading shareds back from the last thread is now a forward-only message only as well.

Register mapping between threads has been changed to conform to the push model. In other words, the consumer thread's dependents are no longer mapped to the producer thread's shareds, but the other way around. This removes the need for the next thread to have terminated before cleaning up a thread. It does make the mapping conceptually a bit more complex, but the implementation is overall simplified.

Delegations have been changed to remove the proxy entry on the creating core. All instructions dealing with a remote place are directly forwarded to that place. Delegations are now also always forced to local creates to avoid issues with group context reservation and exclusive creates.

By having the parent explicitely request all interaction (such as sync and parameter passing) with the child family through its family identifier, families no longer need to know who and where their parent is.

The network interface has been simplified to a message-passing interface for the delegation network, to prepare for an actual message-passing network simulation.

Since globals and shareds are now passed explicitly, long latency operations to registers which are mapped to remote threads would only exist for intra-thread shareds. To simplify the register file and writeback logic, long-latency operations on shareds have been forbidden. This results in all register sends occuring from the pipeline, or the network itself in case of forwards.

Families are now created with pseudo-randomly generated capabilities which are returned in the family identifier and checked for every operation made on families by programs.

Place identifiers now feature a suspend bit, which indicates if the allocate instruction should suspend if no context is available, or return a zero family identifier so the code can perform with special deadlock avoidance operations. Exclusive creates always force the "suspend" bit. The "group" and "default" place IDs have also been separated as the "default" place now inherits the creating thread's family's place, and and the "group" place explicitely requests a group create.

All programs have been rewritten to use the new instruction set as well and a testcase for continuation creates has been added.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3652 e97e5017-a994-416e-809d-76780e9f78db