[mgsim] Fixed COMA and I-Cache bugs
There was a bug in the COMA Cache implementation where it checked the tokens in the line, rather than in the message, causing a possible inconsistency in the different phases of a cycle, causing an assert in the simulation kernel.

In the I-Cache there was a bug where invalidated lines were not considered when memory reads completed, causing those lines to never be filled, causing threads to never execute, causing deadlock.

Both these bugs have been fixed, and a few lines of code cleaned up as well.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3795 e97e5017-a994-416e-809d-76780e9f78db