[mgsim] I-Cache bug fix for invalidated lines
When a cache line is invalidated in the I-Cache and completed from memory, subsequent lookups would still hit it and pretend it was in the loading phase, while it's not. They would suspend on the line never to be woken up, causing deadlock.

This has been fixed by adding a case for a lookup hitting an invalidated line. The action is to stall the lookup. We cannot complete the lookup (even though the data might be there), because that would break memory consistency (i.e. new threads created after a write from another processor causing the invalidation). We cannot allocate a second line with the same tag either. Thus we have to wait with serving the lookup until the current line has returned from memory and its threads serviced.

This change fixes #268.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3856 e97e5017-a994-416e-809d-76780e9f78db