[mgsim] fixed allocate/create/sync register writeback race condition
The register writeback on allocate, create and sync happens via a seperate queue and process from the pipeline. It was possible that the instruction that scheduled the event that causes the writeback does not write back Pending to the register file before the completion writeback is written back. Then, the Pending write back would overwrite the result and the system would eventually deadlock.

To fix this, the separate writeback process must wait with writing back until the target register is Pending or Waiting, which means that the issuing instruction has run through the pipeline.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3916 e97e5017-a994-416e-809d-76780e9f78db