[mgsim] Fixed remote shared/global writes
Remote shared and global writes from the delegation network were assumed to be to local families only, since delegated families are local. However, for asynchronous wakeup between unrelated families, a remote shared write to a group family could also be possible, in which case the remote register has to be forwarded from the delegation network onto the group network (i.e., to the next core, which holds the first thread). Some code had to be added for this check and forwarding.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3931 e97e5017-a994-416e-809d-76780e9f78db