[mgsim] Replaced broken COMA implementation with working COMA implementation.
The original COMA memory system, due to its attempts to be efficient, suffered from several race conditions which were extremely non-trivial to fix. Instead, a whole new COMA protocol has been implemented that, while perhaps not as efficient, at least works. The new protocol uses unidirectional, and not bidirectional, rings. The old COMA tried to keep enough tokens for all caches up-stream that did not have the cache-line. The new COMA does not. Messages grab a token when they can, and get forwarded otherwise. Evictions never go down into directories for injections due to timing problems with administration in the directory.

To support this new protocol, a new kind of arbitration has been created. Previously, only priority-based arbitration existed. This has been extended with a cyclic arbitration, to alternate between clients in case multiple clients want to use a shared resource. This is necessary in the directories to avoid deadlock. The default behavior for all other arbitrators is still the priority-based arbitration.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3942 e97e5017-a994-416e-809d-76780e9f78db