mgsim: changed failure for invalid remote register writes into silent ignore
When writing a remote register that does not exist in the family (for instance, writing a global when the target family has no globals), an assertion is triggered in debug mode and an exception in release mode. This is caused because the code assumes that only valid remote registers will be written. However, this can obviously not always be the case (e.g. optimizing away globals on the callee side), so the network should just ignore those writes.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3974 e97e5017-a994-416e-809d-76780e9f78db