[mgsim] fixed data consistency bug in memory system.
Updates on the COMA ring were not propagated to the L1 caches below an L2 cache, causing data inconsistencies. Furthermore, completed reads are now first merged with pending writes (in a bit of a hack) before being sent to the processors below an L2 cache. The D-Cache also have to be updated to allow for writes to Invalidated lines to be gracefully handled by a stall.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3975 e97e5017-a994-416e-809d-76780e9f78db