[mgsim] Additional delegation network buffer fix
The fix in r4015 introduced another bug, where local allocation responses would bypass the delegation output buffer and want to go back into the delegation input buffer, which is full with the allocation request itself, causing deadlock. The fix is to overwrite the request in the input buffer, like is done for local register responses.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4016 e97e5017-a994-416e-809d-76780e9f78db