[mgsim] Fixed COMA RootDirectory bug.
The problem reported in #276 occured because a read request arrived at the root directory without a line being available for allocation, which is a situation that can occur in rare conditions. The fix was simply to ignore the message and forward it around the ring if no line could be allocated. This way, the request will keep going round until a line can be allocated.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4019 e97e5017-a994-416e-809d-76780e9f78db