[mgsim] Added multiple root directories and DDR channels.
The number of root directories is specified in the configuration with NumRootDirectories. It must be a power of two. The root directories are distributed evenly across the top-level ring. Every root directory has a single DDR3 channel connected to it.

The root directories check the lowest bits of the address after dividing by the line size. In other words, the cache-lines in memory are distributed round-robin across the root directories.

ZLCOMA's topology code had to be adjusted, as it did not properly take all situations into account for NumRootDirectories > 1.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4091 e97e5017-a994-416e-809d-76780e9f78db