[mgsim] Fixed the load-after-load bugs.
Several problems, as indicated by #254, originated from the fact that issuing a load to a register after a prior load to that register, without using the register in between, led to problems.

The fix for this problem lies in exploiting the fact that loads do not use all source operands in the implemented instruction sets. The fix, therefore, makes the destination register also a source operand, specially flagged in the pipeline. When this source operand is read, it is checked for being full, as normal. In this case, however, the source operand can also be empty; hence the flagged situation. If the read stage determines that this source operand is pending, it will wait on it as it would on any pending source operand.

This mechanism ensures that a load will never be issued with a pending destination register.

As a sidenote, on the Sparc, both source operands are used, but since stores require three source operands, support for three source operands is already built into the pipeline. Loads have been altered to use this mechanism as well. This does cause loads to take two cycles instead of one, though.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4189 e97e5017-a994-416e-809d-76780e9f78db