[mgsim] Fixed COMA memory consistency bug.
Sometimes dirty cache-lines are not written back whereas they should be. This is caused by the dirty flag not being copied when a copy is made of a cache-line. When the original, dirty cache-line is then evicted, only the tokens are stored in the root directory. When the non-dirty copy is then evicted, the line is cleared from the system, but not written back, because the line's dirty status has been lost.

This has been fixed by sending the dirty bit along whenever a copy of a cache-line is made, such that the final evict of all copies of a cache-line will always be dirty if the original was.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4193 e97e5017-a994-416e-809d-76780e9f78db