[mgsim] Fixed deadlock bug.
The deadlock was caused by an arbitration issue, where the FPU was prioritized over the D-Cache. The FPU was attempting to 
writeback a value to the register file that was not yet pending. The pending writeback was stuck in the pipeline. The 
pipeline was waiting for the D-Cache to write back a value to the register file.

Because the FPU had priority over the D-Cache, it kept stalling and preventing the D-Cache from completing.

The priorities have been switched.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4194 e97e5017-a994-416e-809d-76780e9f78db