[mgsim-zlcoma-fix] Final fixes to ZLCOMA.
With these fixes, the ZLCOMA has been put in a (mostly) operational state, based on a token counting protocol with write-invalidates and transient tokens. The merge store buffer and evicted line buffer have been removed and several 'optimizations' in the original ZLCOMA have been removed to ensure a robust implementation.

NOTE: although the current implementation passes the tests, it should not be used for production results without additional extensive testing and validating.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/zlcoma-fix@4201 e97e5017-a994-416e-809d-76780e9f78db