[mgsim] Fixed Sparc double-register read bug.
Instructions on the MTSparc that read double registers (like STD) did not properly read their operands. This was caused by the Read Stage, where the shifting of a single register into its place in the operand did not happen correctly. The read value was of the 32-bit integer type, shifted, and or-ed with the 64-bit result. The 32-bit integer should've been cast to 64-bits before shifting.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4216 e97e5017-a994-416e-809d-76780e9f78db