[mgsim-load-balancing] Updated balancing implementation to use single allocates.
A balanced allocate used to send an allocate for a place with size 1 to the best core in the original place. However, this meant that child families with default place stayed on that one core. Since it was desires that such child families still be distributed across the original place, a new allocate type has been introduced: single allocates.

Single allocates allocate only one core in the place, while keeping the place information intact for child families. As a result, some processes needed to be adjusted since such families no longer need to start at a multiple of the place size.

The 'single allocate' has also been made available to programs as flags[1:0] = 3 in the "allocate" instruction. This is useful for single-thread families and dependency families on places larger than 1, to avoid the overhead of allocating and then releasing a context on every core.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/load-balancing@4248 e97e5017-a994-416e-809d-76780e9f78db