[mgsim] Fixed ZLCOMA root directory bug.
A bug in the Root Directory in ZLCOMA could cause a read be sent to external memory for a line that was already being read from external memory if a second read arrived.

The solution is to forward the read, and not send it to external memory, if that line is already being loaded.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4255 e97e5017-a994-416e-809d-76780e9f78db