[mgsim-refactor] Code tree restructuring.
The files have been split into separate 'modules':

sim, the simulator core. This has no knowledge of the architecture.
arch, the components of the simulated architecture.
tools, all tools that work with simulation output or create input.
cli, the command-line version of the simulator.

Note that this commit is only a first move of the files. The makefiles, header guards and includes still need to be update. Some type definitions may have to be split to make the component seperation more strict.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4276 e97e5017-a994-416e-809d-76780e9f78db