[mgsim-refactor] Fixed build process.
The Makefiles have been updated and cleaned and the header includes corrected so that the build works again.

Note that the eight variants of architecture, debug mode and graphics are no longer built at the same time. The three 
options now need to be selected with the configure script. The new options are --disable-assert to disable asserts, and 
--enable-sdl to enable graphics. Furthermore, CXXFLAGS can be used to set or disable the optimization level and debug 
information in the executable. In either way, the executable is always called 'mgsim'.

The idea is that the target (alpha, sparc) will be selected with the configure script as well, á la gcc and binutils.

This change severly simplifies the Makefiles and configure script and is more in line with the autotools philosophy.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4289 e97e5017-a994-416e-809d-76780e9f78db