[mgsim-refactor] Fixed deadlock detection bug.
The family table has no notion of reserved contexts, but for code simplicity it still allocates space for a reserved contexts counter. For deadlock detection, the sum of all free context types is calculated (which should be equal to the total family table size, iff there is no deadlock).

However, the reserved context counter was uninitialized, meaning the sum is non-deterministic, leading to false deadlock reports.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4436 e97e5017-a994-416e-809d-76780e9f78db