[mgsim-refactor] Fixed livelock in COMA.
The livelock was caused by incorrect priorities on the outgoing buffer in the root directory. The forward (the "shortcut" across the flattened ring) was given priority over the responses from memory. Since the top-level ring was full with read requests for cache-lines that were already being read (which are forwarded to be tried again later), the actual responses that would end this cycle never got a chance to be handled.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4437 e97e5017-a994-416e-809d-76780e9f78db