[mgsim-refactor] Implemented storage trace and rewrote memory interface.
To assist in deadlock detection, every process' storage writes have been formalized into a list of 'storage traces'; a contract for a process that it will access storages in one of the predefined ways. The storage traces are verified by the simulator in a debug build.

Storage traces are specified via expression format, which is described in the newly added sim/storagetrace.h.


As a secondary change, the interface with memory has changed. The number of memory clients is no longer assumed in advance to be equal to the number of processors, since there might be non-processor memory clients. The memory interface has been altered to support dynamically adding memory clients. There are two important side effects of this:
* The I-Cache, D-Cache and DCA are now seperate memory clients. Keep this in mind when determining how many clients may be connected to a higher-level memory module (e.g., in COMA, the number of clients per L2 cache goes from 4 to 8).
* The COMA configuration variable NumProcessorsPerL2Cache has been renamed to NumClientsPerL2Cache.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4482 e97e5017-a994-416e-809d-76780e9f78db