[mgsim-refactor] Make delayed I/O reads look more like memory loads.
So far I/O reads would set the register to empty upon issuing, and
back to full upon completion. This would discard any sign extension
information and would not cause waits on thread completion for
outstanding reads. In order to obtain these, the register must be set
to pending with the proper bits set.

git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4505 e97e5017-a994-416e-809d-76780e9f78db