[mgsim] Fixed deadlock in root directories.
The line arbitrator in the COMA and ZLCOMA root directories was a priority-based arbitrator. As a consequence, the coma network was not guaranteed to be able to keep moving. This caused #333: responses got priority, but the entire network was full, except for the requests queue. But the input buffer could not forward into the requests queue because the arbitrator kept awarding access to the responses queue.

It has been solved by changing the priority arbitrator to a round-robin arbitrator, like it was done in the directories a while ago.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4524 e97e5017-a994-416e-809d-76780e9f78db