[mgsim] Fix a consistency bug between L1 and L2.
When a load completes from L2 to L1 while a store is still pending in
the L1 outgoing buffer, the store would not be used to update the line
contents. As a result the L1 line would become outdated and
inconsistent with L2 cache and further loads would miss the new
content from the pending store.

This change fixes the problem by merging incoming lines from L2 to L1
with stores in the outgoing buffer.

git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4576 e97e5017-a994-416e-809d-76780e9f78db